Fix pad mapping
Agana's pads are shifted by half a pad w.r.t. the current mapping.
The current mappgin was based on one of the simulations GetPadPosition
(which was wrong by half a pad). Further inspection shows that there is
another GetPadPosition in the simulation which matches agana; hence I
assume this is the correct behavior.